Add files via upload
This paper aims to assess the effect of Indian States’ fiscal autonomy on women’s empowerment. By using a two-stage least square estimation method with some political factors such as partisan affiliation as an instrument, I found a positive effect of states’ fiscal autonomy on women situation and a negative effect on gender inequality. So, Indian federalism provides better access to education, health care, and economic opportunities for women while men seem to be the main beneficiaries of these policies